requirements.txt
PyGithub == 1.55
requests
